Separated controller from relationType data; response type test
