+++ spec/OpenSkos/RelationType/RelationTypeSpec.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace spec\App\OpenSkos\RelationType;
+
+use PhpSpec\ObjectBehavior;
+
+class RelationTypeSpec extends ObjectBehavior
+{
+    public function it_returns_an_easyrdf_graph()
+    {
+        self::relationTypes()->shouldBeAnInstanceOf('EasyRdf_Graph');
+    }
+}
